Fix: Typo and add oxford comma
Ironically, it was the word "README" that was typo'd in the suggestion of contributing errors in text files. My obsessional neurosis wasn't gonna be able to sleep tonight without fixing it, and then in a sentence about grammar I couldn't help but add an oxford comma. ;-) In my cursory review of the file, I couldn't find any other potential use case as guidance, so I just went with personal stylistic preference. Be well.